Make every eye-config field required; delete the hand-rolled JSON writer
The config codecs used optionalFieldOf throughout, so encode elided any value
equal to its default. EyeConfigJsonWriter existed to defeat that, assembling
JsonObjects by hand — a second copy of the field list that nothing forced you
to update when the codec grew a field.

Nothing read "absent" as different from "default": the loader collapsed both,
and of 503 shipped eyes only crossTarget was ever elided. Switch every field to
fieldOf, migrate those 224 eyes to carry an explicit -1, and let
ConfigFile.CODEC's own encode emit the canonical form.

Fallout:
- EyePlacement stores floats, so 0.22 no longer widens to 0.2199999988079071
  on its way to disk; the writer's round() helper goes with it.
- Its structural filtering and attach-token canonicalization fuse into
  RuntimeConfig.pruned(), shared by both export paths.
- versionRange() moves to VersionRangeMatcher, which owns the range format.
- The reload listener logs the codec's parse error instead of discarding it.
- A malformed export payload is now rejected as malformed rather than decoding
  to an empty config — the service's dead rejection branch is live.
- The round-trip test replaces the writer test's hardcoded field list.

optionalFieldOf stays where absence is real: RuntimeConfigSet's age buckets and
AppearanceOverride's sparse item override.

+++ src/main/java/com/github/crittscott/somegoogly/client/picker/EyeDraft.java
@@ -80,7 +80,8 @@ public class EyeDraft {
     public EyeDefinition toDefinition(int resolvedCrossTarget) {
         EyePlacement placement = new EyePlacement(
                 new Vec3(position[0], position[1], position[2]),
-                eyeScale, irisScale, depth, inclination, azimuth,
+                (float) eyeScale, (float) irisScale, (float) depth,
+                (float) inclination, (float) azimuth,
                 resolvedCrossTarget);
         EyeAppearance appearance = new EyeAppearance(
                 new EyeColor((float) corneaColors[0], (float) corneaColors[1], (float) corneaColors[2]),
